Split SituationInstance out to its own file
